Update comment for AMI00[12]0 override.
The AML is even stupider than always returning 0. It will only return
non-zero for an OS that reports itself as "Windows 2015", at least
on the Threadripper board's AML that I've examined.

Those AMLs also suggest we may need this quirk for AMI0030 as well.
There may be other cases where we need to override the _STA in a
generic way, so we should consider writing code to do that.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@339736 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f