Remove btxld from symlinked host tools
It is only present on amd64/i386 systems which breaks buildworld on
other hosts. In fact there is no need to add it to the bootstrap tools
list since it is already included in the cross-tools phase.
However, for cross-tools it was only built if the host and target
architecture didn't match. After this change it is also built when we
are builtin with a strict $PATH.

Reported By:	mmel


git-svn-id: svn+ssh://svn.freebsd.org/base/head@340176 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f