Complete fix for CID 1007454, CID 1007453: Resource leak in newsyslog
The result of a strdup() was stored in a global variable and not freed
before program exit. This is a follow-up to r343906. That change
attempted to plug these resource leaks but managed to miss a code path
on which the leak still occurs. Plug the leak on that path, too.

MFC after:	3 days
Sponsored by:	Dell EMC Isilon


git-svn-id: svn+ssh://svn.freebsd.org/base/head@344389 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f