arm64: rockchip: clk: rk_clk_composite: Properly use the mask bits
RockChip clocks register have a write mask in the upper 16 bits, if a 1
is present the corresponding bit in the lower 16 ones is set.
Use this instead of always setting the mask to 0xFFFF0000.
This avoids a read of the register.
While here add some debug printf useful for debuging clock problems

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@344576 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f