bridge: Fix STP-related panic
After r345180 we need to have the appropriate vnet context set to delete an
rtnode in bridge_rtnode_destroy().
That's usually the case, but not when it's called by the STP code (through
bstp_notify_rtage()).

We have to set the vnet context in bridge_rtable_expire() just as we do in the
other STP callback bridge_state_change().

Reviewed by:	kevans


git-svn-id: svn+ssh://svn.freebsd.org/base/head@345187 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f