Fix a typo introduced in r344133
The line was misedited to change tt to st instead of
changing ut to st.

The use of st as the denominator in mul64_by_fraction() will lead
to an integer divide fault in the intr proc (the process holding
ithreads) where st will be 0.  This divide by 0 happens after
the total runtime for all ithreads exceeds 76 hours.

Submitted by: bde


git-svn-id: svn+ssh://svn.freebsd.org/base/head@345273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f