Initialize uart_bus_space_mem.
This value was being used uninitialized, resulting in predictable issues
on systems with memory-mapped UART registers.

A case could be made that memmap_bus should be declared in a header
rather than being declared in each .c file which needs to refer to it,
but that's a broader style question.

This commit unbreaks hw.uart.console="mm:..." on ARM64.

Submitted by:	Greg V


git-svn-id: svn+ssh://svn.freebsd.org/base/head@345406 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f