Fix `get_int_via_sysctlbyname(..)` on Jenkins
Initialize `oldlen` to the size of the value, instead of leaving the value
unitialized. Leaving it unitialized seems to work by accident on amd64 when
running 64-bit programs, but not on i386.

This matches patterns in use in other programs.

PR:		237458
Approved by:	emaste (mentor; implicit)
MFC after:	1 week
Tested on:	^/head (amd64), ^/stable/11 (i386)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@346539 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f