Implement uefi_rootdev
If uefi_rootdev is set in the environment, then treat it like a device
path. Convert the string to a device path and see if we can find a
device that matches. If so, use that device at our root dev no matter
what. If it's bad in any way, the boot will fail.

Reviewed by: bcran
Differential Revision: https://reviews.freebsd.org/D20016


git-svn-id: svn+ssh://svn.freebsd.org/base/head@346880 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f