Re-enable loader efi http boot and fix dv_open bug if dv_init failed
The code in efihttp.c was assuming that dv_open wouldn't be called if
dv_init failed. But the dv_init return value is currently ignored.

Add a new variable, `efihttp_init_done` and only proceed in dv_open if
it's true. This fixes the loader on systems without efi http support.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@349404 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f