ObsoleteFiles.inc: remove stale comment
A comment at the top of the file claimed that the file was grouped into
OLD_FILES, OLD_LIBS, then OLD_DIRS, but that hasn't been the case since
the mid-2000s.  Delete the stale comment, add a new comment for the
historical split entries, and move the one more recent entry (from 2013)
to group it into a single logical change.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@355642 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f