keyserv(8): unifdef out __FreeBSD__ and KEYSERV_RANDOM
This doesn't appear to have some active upstream (and it's a steaming pile of
bad 90s crypto design).  Rip out the completely horrible bits and leave the
only mildly less horrible bits.  The whole thing should probably be deleted; to
the extent it purports to provide a security feature: it doesn't.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@355691 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f