Mark hme(4) as deprecated.
It was saved from the initial purge of drivers in fcp-101 due to being
the onboard Ethernet device on a number of sparc64 machines.  Now that
sparc64 is gone, it serves little purpose (PCI cards exist, but are rare
and are unlikely to have been deployed outside Sun systems).

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@357788 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f