virtio: Pass the interrupt type in mmio mode
When we register an interrupt handler we need to pass the intr_type along in
bus_setup_intr().

The interrupt type matters because it is used to decide if we need to enter
NET_EPOCH. That meant that vtmmio-based if_vtnet did not, which led to panics
with INVARIANTS set.

Sponsored by:	Axiado


git-svn-id: svn+ssh://svn.freebsd.org/base/head@358178 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f