vm_radix: prefer __builtin_unreachable() to an unreachable panic()
This provides the needed hint to GCC and offers an annotation for readers to
observe that it's in-fact impossible to hit this point. We'll get hit with a
a -Wswitch error if the enum applicable to the switch above were to get
expanded without the new value(s) being handled.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@358248 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f