[PowerPC] Fix kernel boot on powerpc
Recent changes have caused the vmspace objects to start coming from KVA
instead of direct-mapped memory on powerpc. As far as I can tell, this is
not actually a problem, so we should stop arbitrarily asserting that it is.

I do not know why this was not being triggered before.

Approved by:	jhibbits
Sponsored by:	Tag1 Consulting, Inc.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@361336 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f