Document upgrade procedure in FREEBSD-upgrade
It was pointed out to me that this is the convention for documenting upgrade
instructions, rather than just leaving the instructions in the commit message.
It's possible these commands won't be used again before we transition to git,
but then at least they'll give a path forward for whoever touches this next.

Suggested by:	lwhsu


git-svn-id: svn+ssh://svn.freebsd.org/base/head@361807 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f