Print the arm64 registers in more exception handling panics
It can be useful to get a dump of all registers when investigating why we
received an exception that we are unable to handle. In these cases we
already call panic, however we don't always print the registers.

Add calls to print_registers and print esr and far when applicable.

Sponsored by:	Innovate UK


git-svn-id: svn+ssh://svn.freebsd.org/base/head@363191 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f