Fix a bug introduced by r363001 for the ext_pgs case.
r363001 added support for ext_pgs mbufs to nfsm_uiombuf().
By inspection, I noticed that "mlen" was not set non-zero and, as such, there
would be an iteration of the loop that did nothing.
This patch sets it.
This bug would have no effect on the system, since the ext_pgs mbuf code
is not yet enabled.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@364138 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f