Remove the long obsolete ufm driver.
It was a driver for a USB FM tuner that was available in the market in 2002. I
wrote the driver in 2003. I've not used it since 2005 or so, so it's time to
retire this driver. No userland code ever interfaced to the special device it
created. There's no user base: the last bug I received on this driver was in
2004.

Relnotes: Yes


git-svn-id: svn+ssh://svn.freebsd.org/base/head@364432 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f